moved google connection code to seperate module
