+++ backend/src/google/google.go
@@ -1,4 +1,4 @@
-package main
+package google
 
 import (
 	"context"
