+++ backend/src/main.go
@@ -5,20 +5,24 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	"github.com/gin-gonic/contrib/static"
 )
 
-func setUpRouter() *gin.Engine {
-	r := gin.Default()
-	r.GET("/", func (c *gin.Context) {
+func serveFrontend(router *gin.Engine) {
+	router.Use(static.Serve("/", static.LocalFile("../../frontend/build", true)))
+}	
+
+func setUpRouter(router *gin.Engine) {
+	router.GET("/health", func (c *gin.Context) {
 		c.String(http.StatusOK, "Hello World")
 	})
-	return r
 }
 
 func main() {
-	runGoogleConnection()
-	r := setUpRouter()
-	err := r.Run(":8080");
+	r := gin.Default()
+	setUpRouter(r)
+	serveFrontend(r)
+	err := r.Run(":8080")
 	if err != nil {
 		fmt.Println("Error starting server")
 	}
